+++ doc/benchmark/middleware_params.clj
@@ -22,10 +22,10 @@
  :query-params {"q" "1"},
  :form-params {"q" "1"},
  :path-or-query-params {"q" "1"}}
-;             Execution time mean : 231.132795 ns
-;    Execution time std-deviation : 7.433019 ns
-;   Execution time lower quantile : 222.289139 ns ( 2.5%)
-;   Execution time upper quantile : 240.245797 ns (97.5%)
+;             Execution time mean : 212.587392 ns
+;    Execution time std-deviation : 16.448955 ns
+;   Execution time lower quantile : 198.148389 ns ( 2.5%)
+;   Execution time upper quantile : 231.934333 ns (97.5%)
 
 (-> {:request-method :get :query-string "q=1"}
     (ring-params))
@@ -47,10 +47,10 @@
 (-> {:request-method :get :query-string "q=1"}
     (impl-params)
     :query-params)
-;             Execution time mean : 379.152939 ns
-;    Execution time std-deviation : 36.867174 ns
-;   Execution time lower quantile : 347.568996 ns ( 2.5%)
-;   Execution time upper quantile : 424.896713 ns (97.5%)
+;             Execution time mean : 326.618344 ns
+;    Execution time std-deviation : 23.205939 ns
+;   Execution time lower quantile : 310.829047 ns ( 2.5%)
+;   Execution time upper quantile : 364.956925 ns (97.5%)
 
 (-> {:request-method :get :query-string "q=1"}
     (ring-params)
@@ -72,14 +72,14 @@
 {:request-method :post,
  :query-string "q=1",
  :headers {"content-type" "application/x-www-form-urlencoded"},
- :body-params {"f" "1"},
- :form-params {"f" "1"},
  :query-params {"q" "1"},
- :path-or-query-params {"q" "1"}}
-;             Execution time mean : 532.867349 ns
-;    Execution time std-deviation : 12.994433 ns
-;   Execution time lower quantile : 519.642936 ns ( 2.5%)
-;   Execution time upper quantile : 553.014053 ns (97.5%)
+ :path-or-query-params {"q" "1"},
+ :body-params {"f" "1"},
+ :form-params {"q" "1"}}
+;             Execution time mean : 500.711220 ns
+;    Execution time std-deviation : 28.070796 ns
+;   Execution time lower quantile : 478.613095 ns ( 2.5%)
+;   Execution time upper quantile : 539.363277 ns (97.5%)
 
 (-> {:request-method :post :query-string "q=1"
      :headers {"content-type" "application/x-www-form-urlencoded"}
@@ -107,10 +107,10 @@
      :body (ByteArrayInputStream. (.getBytes "f=1"))}
     (impl-params)
     (select-keys [:query-params :body-params]))
-;             Execution time mean : 1.257514 µs
-;    Execution time std-deviation : 22.734521 ns
-;   Execution time lower quantile : 1.234830 µs ( 2.5%)
-;   Execution time upper quantile : 1.280438 µs (97.5%)
+;             Execution time mean : 1.232769 µs
+;    Execution time std-deviation : 67.237305 ns
+;   Execution time lower quantile : 1.122566 µs ( 2.5%)
+;   Execution time upper quantile : 1.297176 µs (97.5%)
 
 (-> {:request-method :post :query-string "q=1"
      :headers {"content-type" "application/x-www-form-urlencoded"}